+++ sources/popup_ui/js/panels/controller.js
@@ -15,7 +15,6 @@
             ID: "blank",
             TITLE: ""
         },
-        confirmation_dialog: null,
         error:          null,
         get_started:    null,
         main_menu:      null,
@@ -86,7 +85,6 @@
     }
 
     require(["popup/panels/authentication",
-             "popup/panels/confirmation_dialog",
              "popup/panels/error",
              "popup/panels/get_started",
              "popup/panels/main_menu",
@@ -95,12 +93,11 @@
              "popup/panels/success",
              "scripts/interaction/bookmarks_interface",
              "scripts/utilities/dom_manipulation"],
-            (authentication_module, confirmation_dialog_module, error_module, get_started_module,
+            (authentication_module, error_module, get_started_module,
              main_menu_module, on_hold_module, password_setup_module, success_module,
              bookmarks_module, dom_module) =>
             {
                 panels.authentication = authentication_module;
-                panels.confirmation_dialog = confirmation_dialog_module;
                 panels.error = error_module;
                 panels.get_started = get_started_module;
                 panels.main_menu = main_menu_module;